+++ modules/@apostrophecms/schema/lib/addFieldTypes.js
@@ -231,7 +231,9 @@ module.exports = (self) => {
 
   self.addFieldType({
     name: 'checkboxes',
+    dynamicChoices: true,
     async convert(req, field, data, destination) {
+      const choices = await self.getChoices(req, field);
       if (typeof data[field.name] === 'string') {
         data[field.name] = self.apos.launder.string(data[field.name]).split(',');
 
@@ -241,14 +243,14 @@ module.exports = (self) => {
         }
 
         destination[field.name] = _.filter(data[field.name], function (choice) {
-          return _.includes(_.map(field.choices, 'value'), choice);
+          return _.includes(_.map(choices, 'value'), choice);
         });
       } else {
         if (!Array.isArray(data[field.name])) {
           destination[field.name] = [];
         } else {
           destination[field.name] = _.filter(data[field.name], function (choice) {
-            return _.includes(_.map(field.choices, 'value'), choice);
+            return _.includes(_.map(choices, 'value'), choice);
           });
         }
       }
@@ -303,13 +305,9 @@ module.exports = (self) => {
 
   self.addFieldType({
     name: 'select',
+    dynamicChoices: true,
     async convert(req, field, data, destination) {
-      let choices;
-      if ((typeof field.choices) === 'string') {
-        choices = await self.apos.modules[field.moduleName][field.choices](req);
-      } else {
-        choices = field.choices;
-      }
+      const choices = await self.getChoices(req, field);
       destination[field.name] = self.apos.launder.select(data[field.name], choices, field.def);
     },
     index: function (value, field, texts) {